SCP-011: redesign of Scala's collections library.
The Scala collections library underwent a major redesign in Scala 2.8.
While the current design proved to be successful in many areas, several
pain points have become apparent over the years which indicate fundamental
issues with the design that cannot be removed through gradual minor changes.

We propose that the Scala Center work on the implementation of this new design
for the Scala 2.13 standard library.