feature: automate the creation of proposals/README.md
While going over the current status of all the proposals and creating
https://github.com/scalacenter/advisoryboard/issues/85 I figured we
could maintain the proposals/README.md a bit different. I can imagine
it's easy to forget to update that so I think instead it might be a good
idea to instead just use headers in the proposal files themselves. This
way all updates to a proposal and to the current status is just done in
the actual proposal, and then the README can get generated.